edit const
