+++ src/components/Form/PaymentForm/PaymentForm.tsx
@@ -3,9 +3,9 @@ import './PaymentForm.scss';
 
 import { useNavigate } from 'react-router-dom';
 import { Formik, Form } from 'formik';
-import * as Yup from 'yup';
 import Button from '../../Buttons/Button/Button';
 import RadioButtonGroup from '../../FormElements/RadioButtons/RadioButtonGroup/RadioButtonGroup';
+import { isDelayedPublicationOptions } from '../../../utils/fakeData';
 
 export type TFormModel = {
 	isDelayedPublication: boolean;
@@ -13,26 +13,6 @@ export type TFormModel = {
 	publicationTime: Date | null;
 	paymentMethod: number;
 };
-
-type TPropRadioOption = {
-	id: number;
-	booleanValue: boolean;
-	label: string;
-};
-
-const isDelayedPublicationOptions: TPropRadioOption[] = [
-	{
-		id: 0,
-		booleanValue: false,
-		label: 'Опубликовать сейчас',
-	},
-	{
-		id: 1,
-		booleanValue: true,
-		label:
-			'Опубликовать по расписанию',
-	},
-];
 
 const PaymentForm: FC = () => {
 	const fieldNames = {
@@ -51,33 +31,22 @@ const PaymentForm: FC = () => {
 		paymentMethod: 1,
 	};
 
-	const validationSchema = Yup.object();
-	// 	.test(
-	// 		'registerAsSet-required',
-	// 		'Выберите способ оформления сотрудников',
-	// 		validateRegisterAsSet
-	// 	);
-
 	const onSubmit = (values: TFormModel) => {
 		navigate('/success-sent');
 		console.log('Form data', JSON.parse(JSON.stringify(values)));
 	};
 
 	return (
-		<Formik<TFormModel>
-			initialValues={initialValues}
-			validationSchema={validationSchema}
-			onSubmit={onSubmit}
-		>
+		<Formik<TFormModel> initialValues={initialValues} onSubmit={onSubmit}>
 			{() => (
 				<div className="form">
 					<Form>
 						<div className="input-container">
-						<RadioButtonGroup
-									label="Когда вы хотите опубликовать вашу заявку?"
-									name={fieldNames.isDelayedPublication}
-									options={isDelayedPublicationOptions}
-								/>
+							<RadioButtonGroup
+								label="Когда вы хотите опубликовать вашу заявку?"
+								name={fieldNames.isDelayedPublication}
+								options={isDelayedPublicationOptions}
+							/>
 						</div>
 						<div className="two-btn-disposition">
 							<Button
