+++ src/components/Form/TariffForm/TariffForm.tsx
@@ -2,13 +2,13 @@ import { FC } from 'react';
 import './TariffForm.scss';
 import { useNavigate } from 'react-router-dom';
 import { Formik, Form } from 'formik';
-import * as Yup from 'yup';
 import Button from '../../Buttons/Button/Button';
 import NumericInput from '../../FormElements/NumericInput/NumericInput';
 import TariffRadioButtons from '../../FormElements/TariffRadioButtons/TariffRadioButtons';
 import QtyRecruitersRadioButtons from '../../FormElements/QtyRecruitersRadioButtons/QtyRecruitersRadioButtons';
 import FeeRadioButtons from '../../FormElements/FeeRadioButtons/FeeRadioButtons';
 import { calculateMediaSalary } from '../../../utils/calculateMediaSalary';
+import { tariffIdOptions } from '../../../utils/fakeData';
 
 export type TFormModel = {
 	qtyEmployees: number;
@@ -19,51 +19,7 @@ export type TFormModel = {
 	feeName: number | undefined;
 };
 
-type TtariffIdOption = {
-	id: number;
-	name: string;
-	payment_for_employee_start_working: number;
-	payment_for_employee_after_guarantee: number;
-	guarantee_period: number;
-	units_of_measurement_warranty_period: string;
-	description: string;
-};
-
-const tariffIdOptions: TtariffIdOption[] = [
-	{
-		id: 0,
-		name: '100% за выход сотрудника',
-		payment_for_employee_start_working: 100,
-		payment_for_employee_after_guarantee: 0,
-		guarantee_period: 1,
-		units_of_measurement_warranty_period: 'month',
-		description:
-			'Отличный вариант, чтобы на заявку откликнулись «звездные» рекрутеры с опытом.',
-	},
-	{
-		id: 1,
-		name: '50% - после выхода, 50% - после гарантийного периода',
-		payment_for_employee_start_working: 50,
-		payment_for_employee_after_guarantee: 50,
-		guarantee_period: 1,
-		units_of_measurement_warranty_period: 'month',
-		description:
-			'Гарантийный период — испытательный срок. 1 месяц после выхода сотрудника.',
-	},
-	{
-		id: 2,
-		name: '100% - после гарантийного периода',
-		payment_for_employee_start_working: 0,
-		payment_for_employee_after_guarantee: 100,
-		guarantee_period: 1,
-		units_of_measurement_warranty_period: 'month',
-		description:
-			'Это 1 месяц после выхода сотрудника. Доступен выбор одного рекрутера.',
-	},
-];
-
 const options = [1, 2, 3];
-
 const minSalary = 0;
 const maxSalary = 200000;
 
@@ -74,7 +30,7 @@ const TariffForm: FC = () => {
 		tariffId: 'tariffId',
 		qtyRecruiters: 'qtyRecruiters',
 		labelForChoose: 'labelForChoose',
-		feeName: "feeName",
+		feeName: 'feeName',
 	};
 
 	const navigate = useNavigate();
@@ -93,8 +49,8 @@ const TariffForm: FC = () => {
 	type TFeeOption = {
 		id: number;
 		name: string;
-		fee: number | undefined
-	}
+		fee: number | undefined;
+	};
 
 	const feeOptions: TFeeOption[] = [
 		{ id: 0, name: `${medianSalary * 1.5} ₽`, fee: medianSalary * 1.5 },
@@ -102,25 +58,13 @@ const TariffForm: FC = () => {
 		{ id: 2, name: 'Назначить самостоятельно', fee: undefined },
 	];
 
-
-	const validationSchema = Yup.object();
-	// 	.test(
-	// 		'registerAsSet-required',
-	// 		'Выберите способ оформления сотрудников',
-	// 		validateRegisterAsSet
-	// 	);
-
 	const onSubmit = (values: TFormModel) => {
 		navigate('/form/step-5');
 		console.log('Form data', JSON.parse(JSON.stringify(values)));
 	};
 
 	return (
-		<Formik<TFormModel>
-			initialValues={initialValues}
-			validationSchema={validationSchema}
-			onSubmit={onSubmit}
-		>
+		<Formik<TFormModel> initialValues={initialValues} onSubmit={onSubmit}>
 			{() => (
 				<div className="form">
 					<Form>
