+++ src/components/FormElements/DatePicker/DatePicker.scss
@@ -11,18 +11,6 @@
 	border-radius: 8px;
 }
 
-// .date-picker__wrapper {
-//     max-width: 272px;
-//     position: relative;
-// }
-
-// .date-picker__image {
-//     position: absolute;
-//     &_right {
-//         top: 5px;
-//     }
-// }
-
 .react-datepicker {
 	&-wrapper {
 		width: 272px;
