+++ src/components/Popup/Popup.scss
@@ -8,7 +8,6 @@
     width: 100%;
     height: 100%;
     background-color: rgba(0, 0, 0, 0.75);
-    // opacity: 0;
     overflow: hidden;
     overflow-y: auto;
     transition: opacity 0.3s;
